Likelihood rev2: cast to UInt

+++ src/nupic/algorithms/AnomalyLikelihood.cpp
@@ -229,7 +229,7 @@ vector<Real>  AnomalyLikelihood::updateAnomalyLikelihoods(vector<Real> anomalySc
   // Filter the likelihood values. First we prepend the historical likelihoods
   // to the current set. Then we filter the values.  We peel off the likelihoods
   // to return and the last windowSize values to store for later.
-  UInt toCrop = min((unsigned long)this->averagedAnomaly.getMaxWindowSize(), (unsigned long)runningLikelihoods.size());
+  UInt toCrop = min((UInt)this->averagedAnomaly.getMaxWindowSize(), (UInt)runningLikelihoods.size());
   this->runningLikelihoods.insert(runningLikelihoods.end() - toCrop, likelihoods.begin(),likelihoods.end()); //append & crop
   NTA_CHECK(this->runningLikelihoods.size() <= this->averagedAnomaly.getMaxWindowSize());
